+++ admin/includes/view_all_users.php
@@ -67,11 +67,14 @@
 <?php
     //DELETE POST
     if( isset( $_GET['delete'] ) ) {
-        $the_user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = $the_user_id";
-        $delete_query = mysqli_query($connection, $query);
-        confirm_query($delete_query);
-        header('Location: users.php');
+        if( isset($_SESSION['user_role']) && $_SESSION['user_role'] == 'admin') {
+            $the_post_id = mysqli_real_escape_string($connection, $_GET['delete']);
+            $the_user_id = $_GET['delete'];
+            $query = "DELETE FROM users WHERE user_id = $the_user_id";
+            $delete_query = mysqli_query($connection, $query);
+            confirm_query($delete_query);
+            header('Location: users.php');
+        }
     }
